requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic-settings>=2.3.0,<3.0.0
python-multipart>=0.0.9,<1.0.0
huggingface-hub>=0.24.0,<1.0.0
httpx>=0.27.0,<1.0.0
vllm==0.14.0
